requirements.txt
openai
python-dotenv
python-docx
numpy
argparse
fastapi
uvicorn[standard]
lxml
